+++ crates/rils_builtins/src/native_definitions.rs
@@ -12,6 +12,30 @@ pub mod copy {
     rils_stdlib::copy_definition!(decl_rils_trait_metadata);
 }
 
+pub mod default {
+    use rils_builtins_macros::decl_rils_trait_metadata;
+
+    rils_stdlib::default_definition!(decl_rils_trait_metadata);
+}
+
+pub mod eq {
+    use rils_builtins_macros::decl_rils_trait_metadata;
+
+    rils_stdlib::eq_definition!(decl_rils_trait_metadata);
+}
+
+pub mod hash {
+    use rils_builtins_macros::decl_rils_trait_metadata;
+
+    rils_stdlib::hash_definition!(decl_rils_trait_metadata);
+}
+
+pub mod bit_flags {
+    use rils_builtins_macros::decl_rils_trait_metadata;
+
+    rils_stdlib::bitflags_definition!(decl_rils_trait_metadata);
+}
+
 pub mod option {
     use rils_builtins_macros::{decl_rils_metadata, decl_rils_trait_impls};
 
@@ -52,6 +76,10 @@ pub use option::DECLARATION;
 pub const DECLARATIONS: &[crate::BuiltinDeclaration] = &[
     clone::DECLARATION,
     copy::DECLARATION,
+    default::DECLARATION,
+    eq::DECLARATION,
+    hash::DECLARATION,
+    bit_flags::DECLARATION,
     option::DECLARATION,
     result::DECLARATION,
 ];