+++ crates/rils_stdlib/src/stdlib/mod.rs
@@ -1,8 +1,7 @@
-pub mod clone;
-pub mod copy;
 pub mod float;
 pub mod integer;
 pub mod option;
 pub mod prelude;
 pub mod result;
 pub mod string;
+pub mod traits;